Source Code/FrontEnd/requirements.txt
    flask==3.0.3
    torch>=2.1.0
    torchvision>=0.16.0
    timm==1.0.9
    pillow>=10.0.0
    numpy>=1.25.0
    opencv-python>=4.8.0